Fixed specification vocabularies links
To avoid confusion, all the vocabulary file names correspond to the name of the source file in the OASIS GitHub repo. These files are cloned to open-services.net/ns and the .ttl files are converted to produce the RDF/XML and JSON-LD variants that are accessed using HTTP content negotiation.